Add player controller script and game manager object
This commit allows the 2D movement of GameObjects tagged as "Player"
Added to the project should be a "Game Controller", which will hold all
scripts related to general gameplay.
The contents of "Custom Assets" (which were mostly empty placeholders)
were moved up a level to reside in the Assets directory.