Add initial state of integrating Rain AI into project
This version does not function properly, so will need to be fixed before
progress is made. The goal is to have the GameController handle spawning
in a handful of zombie prefabs at the start of the game. The current
behavior tree for the zombies remains untested due to difficulties in
switching logic across the x - z plane instead of the traditional x - y
plane.