Add Skeleton conga line dancers to main branch
This commit contains the SkeletonBehavior tree, which is only to move
towards a GameObject that gets assigned by the Game Manager. A "Player"
skeleton has its controls overridden to do that which the player desires
as the leading zombie in the conga line.

Needed: alteration to player controller to rotate the player object
appropriately via input